requirements.txt
alembic>=1.13,<2.0
pandas>=2.2,<3.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
SQLAlchemy>=2.0,<3.0
streamlit>=1.42,<2.0
